refactor(TableService): 주문테이블의 주문 상태 확인 로직을 주문이 하도록 변경
